chore: optimize memory usage for pie in agg job
extract common methods between snos and agg job into pie module
